Refactoring components into external and internal to show better how dependencies play together.
